requirements.txt
praw==7.7.1
google-generativeai==0.3.1
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
requests==2.31.0
schedule==1.2.0
jinja2==3.1.2
# Auth dependencies
authlib==1.2.1
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.27 